requirements.txt
ollama
streamlit
duckduckgo-search
requests
beautifulsoup4
torch
torchvision
torchaudio
openai-whisper
sounddevice
scipy
numpy
coqui-tts
playsound==1.2.2
pygame
PyPDF2
python-docx
openpyxl
pyautogui
pillow
psutil
pyperclip
secure-smtplib
schedule
pystray
keyboard
python-dotenv
colorama
